Move `array_to_str` to `_utils`
